Moved test to its own kotlin file.
