+++ node/src/integration-test/kotlin/net/corda/node/NodeStartupPerformanceTests.kt
@@ -1,11 +1,7 @@
 package net.corda.node
 
 import com.google.common.base.Stopwatch
-import net.corda.node.logging.logFile
-import net.corda.testing.driver.DriverParameters
 import net.corda.testing.driver.driver
-import net.corda.testing.driver.internal.incrementalPortAllocation
-import org.junit.Assert
 import org.junit.Ignore
 import org.junit.Test
 import java.util.*
@@ -26,26 +22,6 @@ class NodeStartupPerformanceTests {
                 times.add(time)
             }
             println(times.map { it / 1_000_000.0 })
-        }
-    }
-}
-
-class NodeEnvironmentVariablesConversionTests {
-    @Test
-    fun `underscore variable is converted and utilized`() {
-        val portAllocator = incrementalPortAllocation()
-        val sshPort = portAllocator.nextPort()
-
-        driver(DriverParameters(
-                systemProperties = mapOf("corda_sshd_port" to sshPort.toString()),
-                startNodesInProcess = false,
-                portAllocation = portAllocator)) {
-            val hasSsh = startNode().get()
-                    .logFile()
-                    .readLines()
-                    .filter { it.contains("SSH server listening on port") }
-                    .any { it.contains(sshPort.toString()) }
-            assert(hasSsh)
         }
     }
 }
